+++ src/objects/objects.js
@@ -548,6 +548,39 @@ Objects.prototype = {
 				}
 			})
 
+			let _color = null;
+			//[jscastro] added property for wireframes state
+			Object.defineProperty(obj, 'color', {
+				get() { return _color; },
+				set(value) {
+					if (!obj.model || _color === value) return;
+					obj.model.traverse(function (c) {
+						if (c.type == "Mesh" || c.type == "SkinnedMesh") {
+							let materials = [];
+							if (!Array.isArray(c.material)) {
+								materials.push(c.material);
+							} else {
+								materials = c.material;
+							}
+							let m = materials[0];
+							if (value) {
+								c.userData.materials = m;
+								c.material = new THREE.MeshStandardMaterial();
+								c.material.color.setHex(value);
+							} else {
+								c.material.dispose();
+								c.material = c.userData.materials;
+								c.userData.materials.dispose();
+								c.userData.materials = null;
+							}
+
+						}
+					});
+					_color = value;
+				}
+			})
+
+
 			let _selected = false;
 			//[jscastro] added property for selected state
 			Object.defineProperty(obj, 'selected', {
@@ -986,8 +1019,7 @@ Objects.prototype = {
 			text: '',
 			cssClass: 'toolTip text-xs',
 			mapboxStyle: false,
-			topMargin: 0,
-			feature: null
+			topMargin: 0
 		},
 
 		sphere: {